database to dictionary
get the data and convert it to a dictionary (which is almost a json file, and thix can be converted easily to one) and print it on a html page. as a next step we save it as variable, convert it back to a json file and then plot the data as a graph!